Consolidate derivation helper functions into a single file
